fix: Gender를 UserEntity 안으로 이동
