+++ apps/commerce-api/src/main/java/com/loopers/domain/user/UserCommand.java
@@ -3,14 +3,14 @@ package com.loopers.domain.user;
 public interface UserCommand {
     record Create (
             String loginId,
-            Gender gender,
+            UserEntity.Gender gender,
             String birthDate,
             String email
     ) implements UserCommand {
         public static Create of (
                 String loginId, String gender, String birthDate, String email
         ){
-            return new Create(loginId, Gender.from(gender), birthDate, email);
+            return new Create(loginId, UserEntity.Gender.from(gender), birthDate, email);
         }
     }
