+++ apps/commerce-api/src/main/java/com/loopers/domain/user/UserEntity.java
@@ -1,6 +1,8 @@
 package com.loopers.domain.user;
 
 import com.loopers.domain.BaseEntity;
+import com.loopers.support.error.CoreException;
+import com.loopers.support.error.ErrorType;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
@@ -47,5 +49,30 @@ public class UserEntity extends BaseEntity {
     public UserPointVo charge(Long amount) {
         this.point = this.point.charge(amount);
         return this.point;
+    }
+
+    @Getter
+    public enum Gender {
+        MALE("남"),
+        FEMALE("여"),
+        ;
+
+        private final String value;
+
+        Gender(String value) {
+            this.value = value;
+        }
+
+        public static Gender from(String value) {
+            if (value == null || value.isBlank()) {
+                throw new CoreException(ErrorType.BAD_REQUEST, "성별은 비어있을 수 없습니다: " + value);
+            }
+
+            return switch (value) {
+                case "남" -> MALE;
+                case "여" -> FEMALE;
+                default -> throw new CoreException(ErrorType.BAD_REQUEST, "지원하지 않는 성별입니다: " + value);
+            };
+        }
     }
 }
